ONNX import refactored
