+++ include/eddl/serialization/onnx/layers/core/permute_onnx.h
@@ -2,7 +2,20 @@
 #ifndef EDDL_PERMUTE_ONNX_H
 #define EDDL_PERMUTE_ONNX_H
 #include "eddl/serialization/onnx/onnx.pb.h"
+#include "eddl/serialization/onnx/utils_onnx.h"
 #include "eddl/layers/core/layer_core.h"
+
+/*
+ * ONNX IMPORT
+ */
+
+// OPSET: 13, 1
+Layer* build_permute_layer(onnx::NodeProto *node,
+                           map<string, Layer *> &output_node_map,
+                           bool recurrent_net,
+                           LOG_LEVEL log_level,
+                           int dev,
+                           int mem);
 
 /*
  * ONNX EXPORT
