+++ include/eddl/serialization/onnx/layers/core/unsqueeze_onnx.h
@@ -2,7 +2,19 @@
 #ifndef EDDL_UNSQUEEZE_ONNX_H
 #define EDDL_UNSQUEEZE_ONNX_H
 #include "eddl/serialization/onnx/onnx.pb.h"
+#include "eddl/serialization/onnx/utils_onnx.h"
 #include "eddl/layers/core/layer_core.h"
+
+/*
+ * ONNX IMPORT
+ */
+
+// OPSET: 11, 1
+Layer* build_unsqueeze_layer(onnx::NodeProto *node,
+                             map<string, Layer *> &output_node_map,
+                             LOG_LEVEL log_level,
+                             int dev,
+                             int mem);
 
 /*
  * ONNX EXPORT
