+++ include/eddl/serialization/onnx/layers/da/scale_onnx.h
@@ -5,6 +5,17 @@
 #include "eddl/layers/da/layer_da.h"
 
 /*
+ * ONNX IMPORT
+ */
+
+// OPSET: 13
+Layer* build_scale_layer(onnx::NodeProto *node,
+                         map<string, vector<float>> &map_init_values,
+                         map<string, Layer *> &output_node_map,
+                         int dev,
+                         int mem);
+
+/*
  * ONNX EXPORT
  */
